Changed app name in build settings, product name. Slider value is rounded and now seeing value is same as checked.

+++ com.digitalCalc/com.digitalCalc/Classes/CustomSlider.h
@@ -17,6 +17,6 @@
 @property (nonatomic, readonly) CGRect thumbRect;
 
 - (void)initSliderWithValue:(NSInteger)slider_value;
-- (NSInteger)getSliderValue;
+- (int)getSliderValue;
 
 @end
